create frame, dd event, reminder, custom info handlers and convertors
This commit includes the addition of various handler and convertor modules for events, reminders and custom information. The `DaoOperation` trait is implemented in the persistence layer for saving and retrieving entities. The data models for the above entities are also introduced in this commit. They will facilitate necessary conversions between different component needs for getting, saving and deleting events, reminders and custom info. The `ApiEvent`, `ApiReminder`, and `ApiCustomInformation` models were defined with the necessary structure to translate between API and business logic needs. A `FileSystemPersistance` struct and methods for saving, retrieving and deleting entities were also included.

These changes are vital to ensure the proper functioning of events, reminders, and custom info throughout the application. The data models will be used for storing and manipulating data while the handlers will be controlling the logic around it. The convertors will be facilitating necessary transformations between API layer and the core business logic of the application.